feat: submission management

+++ backend/src/main/java/com/quicklance/backend/dto/TaskRequest.java
@@ -1,3 +1,3 @@
 package com.quicklance.backend.dto;
 
-public record TaskRequest(String title, String description, Integer credits) {}
+public record TaskRequest(String title, String description, Long credits) {}
